Extract material-related components and functions
